[DF-1359] Removed false from return instruction
Former-commit-id: 8c1b054b296b244a8d6a308ec67c5408b7bb5302

+++ Sources/Tools/TexturePacker/CommandLineParser.cpp
@@ -213,7 +213,7 @@ String CommandLineParser::GetCommandParamAdditional(const String &command, const
 int32 CommandLineParser::GetCommandsCount()
 {
     Vector<String> & commandLine = Core::Instance()->GetCommandLine();
-    return commandLine.size();
+    return (int32)commandLine.size();
 }
 
 bool CommandLineParser::CheckPosition(int32 commandPosition)
